Update resource mounts to be read-only
Per CIS OpenShift recommendation 1.1.15, the `kubeconfig` should only be
readable (e.g., permissions set to 0600 or more restrictive). Today, the
default permissions for the `kubeconfig` file are 644.

This patch updates the volume mounts to leverage the configMaps instead
of hostPath so we can set the defaultMode to 0600 or decimal
equivalent of 384, by default.